Tweaked config to include doubles for the rates

+++ src/main/java/dungeonmania/movingEntities/Assassin.java
@@ -11,7 +11,7 @@ import dungeonmania.util.Position;
 
 public class Assassin extends Mercenary{
     
-    private int bribeFailRate;
+    private double bribeFailRate;
     private int reconRadius;
     
     /**
@@ -23,7 +23,7 @@ public class Assassin extends Mercenary{
      * @param health
      * @param attack
      */
-    public Assassin(String id, Position position, String type, int health, int attack, int bribe_amount, int bribeFailRate, int reconRadius) {
+    public Assassin(String id, Position position, String type, int health, int attack, int bribe_amount, double bribeFailRate, int reconRadius) {
         super(id, position, type, health, attack, bribe_amount);
         this.bribeFailRate = bribeFailRate;
         this.reconRadius = reconRadius;
